requirements.txt
torchvision>=0.9.1
matplotlib>=3.4.2
easydict>=1.9
torch>=1.8.0
pyyaml>=5.4.1
pillow>=8.2.0
pytorch-lightning